Backend/requirements.txt
Flask==2.2.2
python-dotenv
requests
